requirements.txt
discord.py
python-dotenv
requests
eth-utils